% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact_outbreak.R
\name{replay_outbreak}
\alias{replay_outbreak}
\title{Replays an outbreak up until a specific point in time}
\usage{
replay_outbreak(full_outbreak)
}
\arguments{
\item{full_outbreak}{the full outbreak datasets}
}
\value{
a list with named components having functions (replay/isolate_at) operating on
        the outbreak
}
\description{
Replays an outbreak up until a specific point in time
}
\examples{

set.seed(1235)
SI <- distcrete("gamma", 1L, w = 0, 10, 0.65)
d_contacts <- distcrete("exp", 1L, w=0, 0.05)
x <- contact_outbreak(3, 1.8, SI$r, d_contacts$r, max_cases = 10)
replay <- replay_outbreak(x)
plot(replay$replay(99))
outbreak_2 <- replay$isolate_at(2, 24)
plot(outbreak_2$replay(99))
outbreak_3 <- outbreak_2$isolate_at(3, 25)
plot(outbreak_3$replay(99))

}

